tests: move helper function
